Add condition, if the Neighbour is favorite, you can't deleted him

+++ app/src/main/java/com/openclassrooms/entrevoisins/ui/neighbour_list/NeighbourFragment.java
@@ -2,6 +2,7 @@ package com.openclassrooms.entrevoisins.ui.neighbour_list;
 
 import android.content.Context;
 import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.DividerItemDecoration;
@@ -30,6 +31,7 @@ public class NeighbourFragment extends Fragment {
 
     private NeighbourApiService mApiService;
     private List<Neighbour> mNeighbours;
+    private List<Neighbour> mFavorites;
     private RecyclerView mRecyclerView;
 
 
@@ -64,6 +66,7 @@ public class NeighbourFragment extends Fragment {
      */
     private void initList() {
         mNeighbours = mApiService.getNeighbours();
+        mFavorites = mApiService.getFavorites();
         mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours));
     }
 
@@ -91,8 +94,12 @@ public class NeighbourFragment extends Fragment {
      */
     @Subscribe
     public void onDeleteNeighbour(DeleteNeighbourEvent event) {
-        mApiService.deleteNeighbour(event.neighbour);
-        initList();
+        if (event.neighbour.isFavorite()) {
+            Toast.makeText(this.getContext(),"You can't Delete a Neighbour In Favorite",Toast.LENGTH_LONG).show();
+        } else {
+            mApiService.deleteNeighbour(event.neighbour);
+            initList();
+        }
     }
 
     @Subscribe
